+++ email/src/account/mod.rs
@@ -14,11 +14,14 @@ pub use self::config::CmdsPgpConfig;
 #[cfg(feature = "pgp-gpg")]
 #[doc(inline)]
 pub use self::config::GpgConfig;
+#[cfg(feature = "pgp")]
+#[doc(inline)]
+pub use self::config::PgpConfig;
 #[doc(inline)]
 pub use self::config::{
-    AccountConfig, OAuth2Config, OAuth2Method, OAuth2Scopes, PasswdConfig, PgpConfig,
-    DEFAULT_DRAFTS_FOLDER, DEFAULT_INBOX_FOLDER, DEFAULT_PAGE_SIZE, DEFAULT_SENT_FOLDER,
-    DEFAULT_SIGNATURE_DELIM, DEFAULT_TRASH_FOLDER,
+    AccountConfig, OAuth2Config, OAuth2Method, OAuth2Scopes, PasswdConfig, DEFAULT_DRAFTS_FOLDER,
+    DEFAULT_INBOX_FOLDER, DEFAULT_PAGE_SIZE, DEFAULT_SENT_FOLDER, DEFAULT_SIGNATURE_DELIM,
+    DEFAULT_TRASH_FOLDER,
 };
 #[cfg(feature = "pgp-native")]
 #[doc(inline)]